Completed-Checked80

+++ WE_Project/Controllers/commentsController.cs
@@ -18,34 +18,25 @@ namespace WE_Project.Controllers
         // GET: comments
         public ActionResult Index(int? id)
         {
-
+            if (Session["us"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             var comment = db.comment.Include(c => c.account).Include(c => c.idea).Where(c => c.idea_id == id).OrderBy(c=>c.comment_id);
             return PartialView(comment.ToList());
         }
 
-        // GET: comments/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            comment comment = db.comment.Find(id);
-            if (comment == null)
-            {
-                return HttpNotFound();
-            }
-            return View(comment);
-        }
+       
 
         [HttpPost]
         public JsonResult Create(int ID, string Comment, bool Enonymous)
         {
+
             comment comment = new comment();
             var account_id = Convert.ToInt32(Session["id"]);
             comment.account_id = account_id;
             comment.idea_id = ID;
-            comment.comment_content = Comment;
+            comment.comment_content = Comment.Trim();
             comment.comment_date = DateTime.Now;
             comment.comment_status = Enonymous;
             if (ModelState.IsValid)
@@ -67,66 +58,29 @@ namespace WE_Project.Controllers
             return Json(true);
         }
 
-        // GET: comments/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            comment comment = db.comment.Find(id);
-            if (comment == null)
-            {
-                return HttpNotFound();
-            }
-            ViewBag.account_id = new SelectList(db.account, "account_id", "email", comment.account_id);
-            ViewBag.idea_id = new SelectList(db.idea, "idea_id", "idea_content", comment.idea_id);
-            return View(comment);
-        }
-
-        // POST: comments/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "comment_id,comment_content,account_id,idea_id")] comment comment)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(comment).State = System.Data.Entity.EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            ViewBag.account_id = new SelectList(db.account, "account_id", "email", comment.account_id);
-            ViewBag.idea_id = new SelectList(db.idea, "idea_id", "idea_content", comment.idea_id);
-            return View(comment);
-        }
 
         // GET: comments/Delete/5
-        public ActionResult Delete(int? id)
+        public ActionResult Delete(int? id, int? i)
         {
+            if (Session["us"] == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             comment comment = db.comment.Find(id);
-            if (comment == null)
+            if (comment != null)
             {
-                return HttpNotFound();
+                db.comment.Remove(comment);
+                db.SaveChanges();
+                return RedirectToAction("Details","ideas", new {id = i});
             }
             return View(comment);
         }
 
-        // POST: comments/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
-        {
-            comment comment = db.comment.Find(id);
-            db.comment.Remove(comment);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
+       
 
         protected override void Dispose(bool disposing)
         {
